Fixed an issue with bursted AXI Transactions on the ICache
Size must be set to the number of bytes in a single burst cycle (beat). At the
moment it is set to 0 (unused) in InstructionCache.scala. As a consequence,
when a burst request is made by the cache the slave returns a 32-bit value
repeatedly because the slave thinks that it has passed the wrap boundary.

This fix implements logic for size in the InstructionCache