{I,D}BusSimplePlugin: don't rely on memoryTranslatorPortConfig != null
`null` can be a valid `MemoryTranslator` configuration parameter, for
example in the case of the `PmpPlugin`. Furthermore, the
`IBusCachedPlugin` and `DBusCachedPlugin` do not generally rely on its
`memoryTranslatorPortConfig` to be non-null to instantiate its
translator port, but instead checks whether a `MemoryTranslator`
service exists in the pipeline. Thus it seems best to change
`IBusSimplePlugin` and `DBusSimplePlugin` to behave identically to the
their respective cached versions.  This also prevents accidental CPU
misconfigurations where an MMU or PMP plugin is instantiated and
generally usable, but does not actually take effect in the memory
buses.